change id from int to byte

+++ template_java/src/main/java/cs451/FIFOBroadcast.java
@@ -8,7 +8,7 @@ public class FIFOBroadcast extends URBBroadcast {
     private long[] vc;
     private List<Broadcaster.Message> pending;
 
-    public FIFOBroadcast(List<Host> hosts, int id, long nb_msg) {
+    public FIFOBroadcast(List<Host> hosts, byte id, long nb_msg) {
         super(hosts, id, nb_msg);
         this.vc = new long[hosts.size() + 1];
         for (int i = 0; i < vc.length; i++) {
@@ -21,23 +21,18 @@ public class FIFOBroadcast extends URBBroadcast {
         super.broadcast(msg);
     }
 
-    protected ArrayList<Broadcaster.Message> handleMsg(byte[] msg, int from) {
+    protected ArrayList<Broadcaster.Message> handleMsg(byte[] msg, byte from) {
         ArrayList<Broadcaster.Message> message = super.handleMsg(msg, from);
         ArrayList<Broadcaster.Message> deliveredMessage = new ArrayList<>();
         if (message != null) {
             boolean newMsg = false;
             // check for delivered message
             for (Broadcaster.Message m : message) {
-                // System.out.println("New message : " + m.getMsgId() + " : " + m.getId());
                 if (vc[m.getId()] == m.getMsgId()) {
                     deliveredMessage.add(m);
-                    // System.out.println("Delivering direct : " + m.getMsgId() + " : " +
-                    // m.getId());
                     newMsg = true;
                     vc[m.getId()]++;
                 } else {
-                    // System.out.println("Adding to pending : " + m.getMsgId() + " : " +
-                    // m.getId());
                     pending.add(m);
                 }
             }
@@ -49,11 +44,8 @@ public class FIFOBroadcast extends URBBroadcast {
             while (newMsg) {
                 newMsg = false;
                 for (Broadcaster.Message m : new ArrayList<>(pending)) {
-                    // System.out.println("Checking pending");
                     if (vc[m.getId()] == m.getMsgId()) {
                         deliveredMessage.add(m);
-                        // System.out.println("Delivering from pending : " + m.getMsgId() + " : " +
-                        // m.getId());
                         newMsg = true;
                         pending.remove(m);
                         vc[m.getId()]++;
@@ -61,7 +53,7 @@ public class FIFOBroadcast extends URBBroadcast {
                 }
 
             }
-            // System.out.println("Delivering : " + deliveredMessage.size());
+
             if (deliveredMessage.size() == 0) {
                 return null;
             }
